refactor(aipanel): 拆出 useAIChatStreamEvents 流式事件 hook
- 新增 components/ai/useAIChatStreamEvents.ts：ai-chat-stream 流式事件总 reducer（runtime_phase/collaboration_*/assistant_*/tool_*/delta/done/error/cancelled）、协同 pending 卡片确认与协同锁自动上锁三个 effect，闭包依赖经 deps 同名注入，代码零改动
- AIPanel.tsx 5963 -> 4692 行
